Implement back-button + lots of fixes
- use vue-router to navigate between landing page and application
- create constants module as centralized constants safe
- create DataHelper module that will be used to fetch data centrally
- change some styles that looked odd
