+++ frontend/js/components/pages/Wahlomat.vue
@@ -1,242 +1,77 @@
 <script>
-import VueSwing from "vue-swing";
-import { setTimeout } from "timers";
+// import subpage components
+import ThesisSwiper from "./ThesisSwiper";
+import ResultPage from "./ResultPage";
 
-// Import Icon components
-import ThumbsDownIcon from "../icons/ThumbsDown";
-import ThumbsUpIcon from "../icons/ThumbsUp";
-import NeutralIcon from "../icons/Neutral";
-
-// Import element components
-import ThesisCard from "../elements/ThesisCard";
-
-// Import helper classes
-import ThesisStack from "../../helpers/ThesisStack";
-import * as SwipeStackHelper from "../../helpers/SwipeStackHelper";
-import * as ResultCalculator from "../../helpers/ResultCalculator";
+// import necessary constants
+import { RESULTPAGE, THESISSWIPER } from "../../helpers/Constants"
 
 export default {
   components: {
-    VueSwing,
-    ThumbsDownIcon,
-    ThumbsUpIcon,
-    NeutralIcon,
-    ThesisCard
+    ThesisSwiper,
+    ResultPage
   },
-  props: ["election"],
   data() {
     return {
-      swingConfig: {
-        allowedDirections: [
-          VueSwing.Direction.LEFT,
-          VueSwing.Direction.RIGHT,
-          VueSwing.Direction.UP,
-          VueSwing.Direction.DOWN
-        ],
-        throwOutConfidence: (xOffset, yOffset, element) => {
-          let xConfidence = Math.min(
-            Math.abs(xOffset) / element.offsetWidth,
-            1
-          );
-          let yConfidence = Math.min(
-            Math.abs(yOffset) / element.offsetHeight,
-            1
-          );
-
-          return Math.min(Math.max(xConfidence, yConfidence) * 2, 1);
-        }
-      },
-      currentSwipeDirection: null,
-      thesisStack: new ThesisStack(this.election.thesen)
+    displayedComponent: THESISSWIPER,
+      activeElection: null,
+      activeElectionResult: null,
+      pages: { resultPage: RESULTPAGE, thesisSwiper: THESISSWIPER }
     };
   },
   methods: {
-    getDirectionFromSymbol(symbol) {
-      return String(symbol)
-        .replace("Symbol", "")
-        .replace("(", "")
-        .replace(")", "")
-        .toLowerCase();
+    switchPage(page) {
+      this.displayedComponent = page;
     },
-    answer(thesisId, answer) {
-      this.thesisStack.answer(thesisId, answer);
-
-      // If all cards have been answered, emit an event with the results as payload
-      // so that App.vue can handle page switching to the results page
-      if (this.thesisStack.Finished) {
-        this.$emit(
-          "finished",
-          ResultCalculator.getResults(this.thesisStack.answeredTheses)
-        );
-      }
+    startWahlomat(election) {
+      this.activeElection = election;
+      this.switchPage("");
     },
-    goBack() {
-      if (this.thesisStack.AnsweredTheses.length > 0) {
-        let previousThesis = this.thesisStack.AnsweredTheses[
-          this.thesisStack.answeredTheses.length - 1
-        ];
+    showResults(results) {
+      // Add the list info text to the result objects, because the result page needs it to display it
+      // TODO: replace dummy text with real info once a backend route for list infos exists
+      this.activeElectionResult = results.map(x => {
+        x.info =
+          "Morbi eleifend tellus ac leo sodales, dictum sagittis nisi tincidunt. Curabitur ut laoreet enim. Proin porta condimentum nulla ac tempor. Suspendisse vel ante diam. Fusce posuere, justo nec rutrum ultricies, enim urna fringilla dolor, id varius tellus libero semper nunc. Cras non dui elementum, suscipit quam et, vehicula justo. Pellentesque rutrum vestibulum dolor in finibus. Integer tempor scelerisque mollis. Ut eget venenatis nisl. Proin tristique ipsum eget felis condimentum feugiat.";
+        return x;
+      });
 
-        let previousCard = this.$refs["card" + previousThesis.thesis._id][0]
-          .$el;
-
-        let poppedThesis = this.thesisStack.revert();
-
-        let swingElement = this.$refs.vueswing.cards[
-          this.thesisStack.theses.length -
-            1 -
-            this.thesisStack.answeredTheses.length
-        ];
-
-        let resultObject = SwipeStackHelper.getResultObjectFromString(
-          poppedThesis.result
-        );
-
-        SwipeStackHelper.swipeInCard(
-          previousCard,
-          SwipeStackHelper.getResultObjectFromString(previousThesis.result)
-        );
-
-        swingElement.throwIn(
-          resultObject.xCoordinate,
-          resultObject.yCoordinate
-        );
-      }
-    },
-    buttonClicked(answerString) {
-      let answer = SwipeStackHelper.getResultObjectFromString(answerString);
-
-      // Get the actual div that makes up the card to change its style during the animation...
-      let activeCard = this.$refs["card" + this.thesisStack.ActiveThesis._id][0]
-        .$el;
-
-      // ...and animate a swipe-out in the direction that fits the provided answer
-      SwipeStackHelper.swipeOutCard(activeCard, answer);
-
-      // The throwout animation will take 40 * 5 milliseconds, so we execute this after that time has passed
-      setTimeout(() => {
-        // Get the element that represents the swiped card internally for the swing stack
-        let swingElement = this.$refs.vueswing.cards[
-          this.thesisStack.theses.length -
-            1 -
-            this.thesisStack.answeredTheses.length
-        ];
-
-        // Use swing's built-in method to "officially" swipe out the card so that swing realizes it
-        // This will also trigger the regular swipeout event of the direction corresponding to the coordinates
-        swingElement.throwOut(answer.xCoordinate, answer.yCoordinate);
-
-        // Reset the 'left' and 'top' properties of the card style to undefined because they were only needed for
-        // the swipeout animation and would otherwise interfere with the throwin animation when going back a card
-        SwipeStackHelper.resetCardPosition(activeCard);
-      }, 40 * 5);
-    },
-    dragging(card, throwDirection, throwOutConfidence) {
-      let direction = this.getDirectionFromSymbol(throwDirection);
-      this.currentSwipeDirection = direction;
-
-      SwipeStackHelper.dragging(card, direction, throwOutConfidence);
-    },
-    finishedDragging(card) {
-      SwipeStackHelper.finishedDragging(card);
-
-      this.currentSwipeDirection = null;
+      this.switchPage(this.pages.resultPage);
     }
   }
 };
 </script>
 <template>
-  <div class="site-container">
-    <div>
-      <h1 class="page-headline">{{election.name.toUpperCase()}}</h1>
-      <vue-swing
-        @throwoutleft="(payload) => answer(payload.target.id, 'no')"
-        @throwoutright="(payload) => answer(payload.target.id, 'yes')"
-        @throwoutdown="(payload) => answer(payload.target.id, 'neutral')"
-        @throwoutup="(payload) => answer(payload.target.id, 'skip')"
-        @dragmove="(payload) => this.dragging(payload.target, payload.throwDirection, payload.throwOutConfidence)"
-        @dragend="(payload) => this.finishedDragging(payload.target)"
-        :config="swingConfig"
-        ref="vueswing"
-      >
-        <ThesisCard
-          v-for="thesis in election.thesen"
-          v-show="!thesisStack.AnsweredTheses.some((x) => x.thesis._id === thesis._id)"
-          :key="thesis._id"
-          :id="thesis._id"
-          :thesis="thesis"
-          :distanceFromActiveThesis="election.thesen.indexOf(thesisStack.ActiveThesis)-election.thesen.indexOf(thesis)"
-          :index="election.thesen.length - election.thesen.indexOf(thesis)"
-          :totalThesesCount="election.thesen.length"
-          :currentSwipeDirection="currentSwipeDirection"
-          :ref="'card' + thesis._id"
-        />
-      </vue-swing>
-    </div>
-    <div class="button-container">
-      <div>
-        <button @click="buttonClicked('no')" class="no circleButton">
-          <ThumbsDownIcon
-            width="'27px'"
-            :height="'27px'"
-            :viewBox="'0 0 24 24'"
-            :fillColor="'#fff'"
-          />
-        </button>
-        <button
-          @click="buttonClicked('neutral')"
-          class="neutral circleButton"
-          style="width:3rem;height:3rem;"
-        >
-          <NeutralIcon :width="'24px'" :height="'24px'" :viewBox="'0 0 24 24'" :fillColor="'#333'"/>
-        </button>
-        <button @click="buttonClicked('yes')" class="yes circleButton">
-          <ThumbsUpIcon
-            :width="'27px'"
-            :height="'27px'"
-            :viewBox="'0 0 24 24'"
-            :fillColor="'#fff'"
-          />
-        </button>
-      </div>
-      <a class="link-button" style="margin:15px;" @click="goBack">Frage zurück</a>
-      <a class="link-button" @click="buttonClicked('skip')">Frage überspringen</a>
-    </div>
-  </div>
+  <ThesisSwiper
+    v-if="displayedComponent === pages.thesisSwiper"
+    :election="activeElection"
+    @finished="showResults"
+  ></ThesisSwiper>
+  <ResultPage v-else-if="displayedComponent === pages.resultPage" :results="activeElectionResult"></ResultPage>
 </template>
-<style scoped>
-.button-container {
-  display: flex;
-  flex-direction: column;
-  align-content: center;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-  top: calc(60% + 50px);
-  width: 100%;
+<style>
+body {
+  background-color: #262626;
+  margin: 0;
+  font-family: Helvetica, sans-serif;
 }
 
-.circleButton {
-  cursor: pointer;
-  border: 0;
-  height: 4rem;
-  width: 4rem;
-  border-radius: 3rem;
-  margin: 1em;
-  transition: all 0.1s ease;
-  box-shadow: 0 0.75rem 30px 0 rgba(0, 0, 0, 0.4);
+.yes {
+  background: linear-gradient(135deg, #768c54 0, #4e5c37);
 }
 
-.button-container > button:hover {
-  box-shadow: 0 0.75rem 30px 0 rgba(0, 0, 0, 0.7);
+.neutral {
+  background: linear-gradient(135deg, #c1c1c1 0, #9b9b9b);
 }
 
-.link-button {
-  text-decoration: underline;
-  color: #aaa;
-  cursor: pointer;
+.no {
+  background: linear-gradient(135deg, #a63232 0, #6b2020);
 }
 
-.site-container {
-  height: 100vh;
+.page-headline {
+  text-align: center;
+  margin: 2rem;
+  font-size: 25px;
+  color: #fff;
 }
 </style>
