+++ frontend/js/helpers/Constants.js
@@ -1,0 +1,31 @@
+export const YES = {
+  answer: 'Ja',
+  direction: 'left',
+  xCoordinate: 90,
+  yCoordinate: 0
+}
+
+export const NO = {
+  answer: 'Nein',
+  direction: 'right',
+  xCoordinate: -90,
+  yCoordinate: 0
+}
+
+export const NEUTRAL = {
+  answer: 'Neutral',
+  direction: 'down',
+  xCoordinate: 0,
+  yCoordinate: 90
+}
+
+export const SKIP = {
+  answer: 'Überspringen',
+  direction: 'up',
+  xCoordinate: 0,
+  yCoordinate: -90
+}
+
+export const RESULTPAGE = 'resultpage'
+
+export const THESISSWIPER = 'thesisswiper'
