--- frontend/js/helpers/PageManager.js
@@ -1,3 +1,0 @@
-export const WAHLOMAT = "wahlomat"
-export const LANDINGPAGE = "landingpage"
-export const RESULTPAGE = "resultpage"
