+++ frontend/js/helpers/SwipeStackHelper.js
@@ -1,7 +1,19 @@
-/* eslint-disable operator-linebreak */
-/* eslint-disable space-before-function-paren */
-/* eslint-disable indent */
-export function swipeOutCard(card, direction) {
+import {
+  YES,
+  NO,
+  NEUTRAL,
+  SKIP
+} from './Constants'
+
+function getStringFromSymbol (symbol) {
+  return String(symbol)
+    .replace('Symbol', '')
+    .replace('(', '')
+    .replace(')', '')
+    .toLowerCase()
+}
+
+export function swipeOutCard (card, direction) {
   if (direction === NEUTRAL || direction === SKIP) {
     for (let i = 0; i < 40; i++) {
       setTimeout(() => {
@@ -25,7 +37,7 @@ export function swipeOutCard(card, direction) {
   }
 }
 
-export function swipeInCard(card, direction) {
+export function swipeInCard (card, direction) {
   if (direction === NEUTRAL || direction === SKIP) {
     for (let i = 40; i >= 0; i--) {
       setTimeout(() => {
@@ -49,59 +61,27 @@ export function swipeInCard(card, direction) {
   }
 }
 
-export const YES = {
-  answer: 'yes',
-  direction: 'left',
-  xCoordinate: 90,
-  yCoordinate: 0
-}
-
-export const NO = {
-  answer: 'no',
-  direction: 'right',
-  xCoordinate: -90,
-  yCoordinate: 0
-}
-
-export const NEUTRAL = {
-  answer: 'neutral',
-  direction: 'down',
-  xCoordinate: 0,
-  yCoordinate: 90
-}
-
-export const SKIP = {
-  answer: 'skip',
-  direction: 'up',
-  xCoordinate: 0,
-  yCoordinate: -90
-}
-
-export function getResultObjectFromString(searchString) {
-  return [this.YES, this.NO, this.NEUTRAL, this.SKIP].filter(x => x.answer === searchString || x.direction === searchString)[0]
-}
-
-export function dragging(card, direction, throwOutConfidence) {
+export function dragging (card, direction, throwOutConfidence) {
   // Let opacity decrease when the card gains throwOutConfidence but don't let it fall lower than 0.3
   card.style.opacity = Math.max(1 - throwOutConfidence, 0.7)
 
   if (throwOutConfidence > 0.1) {
-    if (direction === 'right') {
+    if (direction === NO) {
       card.style.background =
         'hsl(83.6,25%,' +
         Math.min(43.9 + (1 - throwOutConfidence) * 100, 70) +
         '%)'
-    } else if (direction === 'left') {
+    } else if (direction === YES) {
       card.style.background =
         'hsl(0,53.7%,' +
         Math.min(42.4 + (1 - throwOutConfidence) * 100, 70) +
         '%)'
-    } else if (direction === 'down') {
+    } else if (direction === NEUTRAL) {
       card.style.background =
         'hsl(0,0%,' +
         Math.min(75.7 + (1 - throwOutConfidence) * 100, 70) +
         '%)'
-    } else if (direction === 'up') {
+    } else if (direction === SKIP) {
       card.style.background =
         'hsl(0,0%,' +
         Math.min(75.7 + (1 - throwOutConfidence) * 100, 70) +
@@ -110,12 +90,20 @@ export function dragging(card, direction, throwOutConfidence) {
   }
 }
 
-export function resetCardPosition(card) {
+export function resetCardPosition (card) {
   card.style.left = undefined
   card.style.top = undefined
 }
 
-export function finishedDragging(card) {
+export function finishedDragging (card) {
   card.style.opacity = 1
   card.style.background = '#fff'
+}
+
+export function getResultObjectFromValue (value) {
+  return [YES, NO, NEUTRAL, SKIP].filter(x =>
+    x.answer === value ||
+    x.direction === value ||
+    x.answer === getStringFromSymbol(value) ||
+    x.direction === getStringFromSymbol(value))[0]
 }
