+++ package.json
@@ -32,21 +32,22 @@
     "cors": "^2.8.5",
     "express": "^4.16.4",
     "express-http-proxy": "^1.5.1",
-    "fs": "0.0.1-security",
-    "laravel-mix": "^4.0.15",
-    "generate-password": "^1.4.1",
-    "mongoose": "^5.5.7",
-    "swing": "^3.1.4",
-    "vue": "^2.6.10",
-    "vue-swing": "0.0.10",
     "express-session": "^1.16.1",
+    "fs": "0.0.1-security",
+    "generate-password": "^1.4.1",
+    "laravel-mix": "^4.0.15",
+    "mongoose": "^5.5.7",
     "mongoose-type-email": "^1.0.11",
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
     "passport-local-mongoose": "^5.0.1",
     "sendmail": "^1.4.1",
     "session-file-store": "^1.2.0",
-    "uuid": "^3.3.2"
+    "swing": "^3.1.4",
+    "uuid": "^3.3.2",
+    "vue": "^2.6.10",
+    "vue-router": "^3.0.6",
+    "vue-swing": "0.0.10"
   },
   "devDependencies": {
     "browser-sync": "^2.26.5",
